pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "psychotropic"
description = "A Discord bot built for harm reduction and chemistry."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
version = "0.0.1"
keywords = []
authors = [
  { name = "xyzt_", email = "xyzt@psychonaut.fr" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "babel~=2.17",
  "discord-py~=2.0",
  "httpx~=0.23",
  "mistune~=3.0",
  "Pillow~=9.2",
  "rdkit~=2024.9",
  "tomli~=2.3", # Required for babel to read pyproject.toml on py < 3.11
]

[dependency-groups]
dev = [
  "ruff>=0.14.10",
  "ty>=0.0.17",
]

[project.urls]
Documentation = "https://github.com/x-yzt/psychotropic#readme"
Issues = "https://github.com/x-yzt/psychotropic/issues"
Source = "https://github.com/x-yzt/psychotropic"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM"]
# Ignore % and f-strings interpolations upgrade. And I know how itertool works, thanks.
ignore = ["UP032", "UP031", "B905"]

[tool.ty.rules]
invalid-method-override = "warn"

[[tool.ty.overrides]]
include = ["psychotropic/cogs/games/**"]

[tool.ty.overrides.rules]
unresolved-attribute = "warn"

[tool.ty.analysis]
allowed-unresolved-imports = ["psychotropic.providers.**"]

[tool.babel]

[[tool.babel.mappings]]
method = "python"
pattern = "**.py"
encoding = "utf-8"
# For now this is still worked out at Babel, so it needs to be explicitely listed in -k CLI args
keywords = ["_", "localize", "localize_fmt"]
